google-appengine/requirements.txt
# This requirements file lists all third-party dependencies for this project.
#
# Run 'pip install -r requirements.txt -t lib/' to install these dependencies
# in `lib/` subdirectory.
#
# Note: The `lib` directory is added to `sys.path` by `appengine_config.py`.
Flask==1.1.1
google-cloud-ndb==0.1.0
requests==2.22.0
Flask-Login==0.4.1
oauthlib==3.0.1
